Refactoring okhttp jmx metric code.
