+++ coreservices/partsrelationshipservice/connector/prs-connector-commons/src/test/java/net/catenax/prs/connector/util/JsonUtilTest.java
@@ -1,17 +1,14 @@
 package net.catenax.prs.connector.util;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.eclipse.dataspaceconnector.monitor.ConsoleMonitor;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.mockingDetails;
 import static org.mockito.Mockito.when;
 
 class JsonUtilTest {
