+++ coreservices/partsrelationshipservice/connector/prs-connector-provider/src/main/java/net/catenax/prs/connector/provider/PartsRelationshipServiceApiExtension.java
@@ -9,12 +9,14 @@
 //
 package net.catenax.prs.connector.provider;
 
+import io.micrometer.jmx.JmxMeterRegistry;
 import net.catenax.prs.client.ApiClient;
 import net.catenax.prs.client.api.PartsRelationshipServiceApi;
 import net.catenax.prs.connector.annotations.ExcludeFromCodeCoverageGeneratedReport;
-import net.catenax.prs.connector.metrics.OkHttpClientProvider;
-import okhttp3.OkHttpClient;
+import net.catenax.prs.connector.http.OkHttpClientProvider;
+import net.catenax.prs.connector.metrics.MeterRegistryProvider;
 import net.catenax.prs.connector.util.JsonUtil;
+import okhttp3.OkHttpClient;
 import org.eclipse.dataspaceconnector.policy.model.Action;
 import org.eclipse.dataspaceconnector.policy.model.AtomicConstraint;
 import org.eclipse.dataspaceconnector.policy.model.LiteralExpression;
@@ -54,13 +56,20 @@ public class PartsRelationshipServiceApiExtension implements ServiceExtension {
      */
     @Override
     public void initialize(final ServiceExtensionContext context) {
+        /*
+          Register JmxMeterRegistry in global context for re-use.
+         */
+        context.registerService(JmxMeterRegistry.class, MeterRegistryProvider.jmxMeterRegistry());
+        
+        final var httpClient = OkHttpClientProvider.httpClient(context.getService(JmxMeterRegistry.class));
+        /*
+            Overrides edc core OkHttpClient to expose micrometer metrics.
+         */
+        context.registerService(OkHttpClient.class, httpClient);
+
         final var monitor = context.getMonitor();
-
-        final OkHttpClient client = OkHttpClientProvider.httpClient();
-        context.registerService(OkHttpClient.class, client);
-
         final var prsApiUrl = context.getSetting("PRS_API_URL", "http://localhost:8080");
-        final var prsClient = new PartsRelationshipServiceApi(new ApiClient(client));
+        final var prsClient = new PartsRelationshipServiceApi(new ApiClient(httpClient));
         prsClient.getApiClient().setBasePath(prsApiUrl);
 
         final var jsonUtil = new JsonUtil(monitor);
