+++ coreservices/partsrelationshipservice/integration-tests/src/test/java/net/catenax/prs/systemtest/ConnectorSystemTests.java
@@ -22,7 +22,6 @@ import java.net.http.HttpResponse;
 import java.time.Duration;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.UUID;
 
 import static io.restassured.RestAssured.given;
 import static java.lang.String.format;
